+++ sdk/core/azure-core/inc/azure/core/credentials/token_credential_options.hpp
@@ -3,7 +3,7 @@
 
 /**
  * @file
- * @brief Options for #Azure::Core::Credentials::TokenCredential.
+ * @brief Client options for #Azure::Core::Credentials::TokenCredential.
  */
 
 #pragma once
@@ -12,7 +12,7 @@
 
 namespace Azure { namespace Core { namespace Credentials {
   /**
-   * @brief Defines options for #Azure::Core::Credentials::TokenCredential.
+   * @brief Client options for #Azure::Core::Credentials::TokenCredential.
    *
    */
   struct TokenCredentialOptions : public Azure::Core::_internal::ClientOptions