+++ sdk/core/azure-core/test/ut/client_options.cpp
@@ -28,11 +28,13 @@ struct FakeTransport final : public HttpTransport
 
 struct PerCallPolicy final : public HttpPolicy
 {
-  std::unique_ptr<RawResponse> Send(Request& request, NextHttpPolicy policy, Context const& context)
-      const override
+  std::unique_ptr<RawResponse> Send(
+      Request& request,
+      NextHttpPolicy nextPolicy,
+      Context const& context) const override
   {
     (void)request;
-    (void)policy;
+    (void)nextPolicy;
     (void)context;
     return std::make_unique<RawResponse>(3, 3, HttpStatusCode::Gone, "IamAPerCallPolicy");
   }
@@ -45,11 +47,13 @@ struct PerCallPolicy final : public HttpPolicy
 
 struct PerRetryPolicy final : public HttpPolicy
 {
-  std::unique_ptr<RawResponse> Send(Request& request, NextHttpPolicy policy, Context const& context)
-      const override
+  std::unique_ptr<RawResponse> Send(
+      Request& request,
+      NextHttpPolicy nextPolicy,
+      Context const& context) const override
   {
     (void)request;
-    (void)policy;
+    (void)nextPolicy;
     (void)context;
     return std::make_unique<RawResponse>(6, 6, HttpStatusCode::ResetContent, "IamAPerRetryPolicy");
   }