+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_options.hpp
@@ -638,7 +638,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
       /**
        * The maximum number of threads that may be used in a parallel transfer.
        */
-      int Concurrency = 5;
+      int32_t Concurrency = 5;
     } TransferOptions;
   };
 
@@ -689,7 +689,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
       /**
        * The maximum number of threads that may be used in a parallel transfer.
        */
-      int Concurrency = 5;
+      int32_t Concurrency = 5;
     } TransferOptions;
   };